webshot2: Take Screenshots of Web Pages

Takes screenshots of web pages, including Shiny applications and R Markdown documents. 'webshot2' uses headless Chrome or Chromium as the browser back-end.

See also

Author

Maintainer: Winston Chang winston@posit.co

Other contributors:

Barret Schloerke barret@posit.co (ORCID) [contributor]

Posit Software, PBC (03wc8by49) [copyright holder, funder]